-GlobalSetting -> value converted to uppercase enum -fix Login response -> utf8 now -moderation posts list sorted by time asc -add isolation level preventing concurrent post modification -add post dynamic update -fix post text wrapping adjust some posts creation time -add swagger login and fix another bugs...

+++ blog-engine-impl/src/main/java/com/skillbox/blog/dto/response/ResponseAllPostsDto.java
@@ -1,11 +1,10 @@
 package com.skillbox.blog.dto.response;
 
+import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.List;
 
 @Data
 @Builder
@@ -13,6 +12,6 @@ import java.util.List;
 @NoArgsConstructor
 public class ResponseAllPostsDto {
 
-  int count;
+  long count;
   List<PartInfoOfPosts> posts;
 }
